LPS-44354 Merge LocalizedSettings and TypedSettings
Having them separate ends up creating confusion.
